requirements.txt
pandas==2.2.2
gensim==4.3.2
nltk==3.8.1
python-dotenv==1.0.1
matplotlib==3.8.4
ipykernel==6.29.4
jupyter==1.0.0
spacy==3.7.5
wordcloud==1.9.3
pyLDAvis==3.4.1
bertopic==0.16.2
matplotlib==3.8.4
scikit-learn==1.4.2
numpy==1.26.4
sentence-transformers==3.0.1
datasets==2.20.0
transformers<=4.41.2
accelerate==0.31.0
torch==2.3.1
huggingface-hub==0.23.4
seaborn==0.13.2
sentencepiece==0.2.0
imbalanced-learn==0.12.3
nlpaug==1.1.11
googletrans==4.0.0-rc1
streamlit==1.38.0
mlflow==2.15.1